move more selection logit out of code mirror and engine connection
